refactor(E2eiEnrollment): move `get_refresh_token` into module tree
